+++ segno/__init__.py
@@ -645,6 +645,20 @@ class QRCode:
         =============    ==============================================================
 
 
+        **X PixMap (XPM)**
+
+        =============    ==============================================================
+        Name             Description
+        =============    ==============================================================
+        kind             "xpm"
+        scale            integer
+        color            Default: "#000" (black).
+        background       Default value ``#fff`` (white)
+                         ``None`` indicates a transparent background.
+        name             Name of the variable (default: "img")
+        =============    ==============================================================
+
+
         :param out: A filename or a writable file-like object with a
                 ``name`` attribute. Use the `kind` parameter if `out` is
                 a :py:class:`io.ByteIO` or :py:class:`io.StringIO` stream which